Move IServer and IInstance to TGS.Interface.Wrappers
