+++ TGS.Interface/TGS.Interface.csproj
@@ -65,9 +65,9 @@
     <Compile Include="Definitions.cs" />
     <Compile Include="Enumerations.cs" />
     <Compile Include="Helpers.cs" />
-    <Compile Include="IInstance.cs" />
+    <Compile Include="Wrappers\IInstance.cs" />
     <Compile Include="InstanceMetadata.cs" />
-    <Compile Include="IServer.cs" />
+    <Compile Include="Wrappers\IServer.cs" />
     <Compile Include="IServerInterface.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Components\Repository.cs" />
